+++ src/main/java/eu/gir/girsignals/blocks/SignalBUELight.java
@@ -2,6 +2,7 @@ package eu.gir.girsignals.blocks;
 
 import eu.gir.girsignals.EnumSignals.BUE_LIGHT;
 import eu.gir.girsignals.SEProperty;
+import eu.gir.girsignals.SEProperty.ChangeableStage;
 import eu.gir.girsignals.init.GIRItems;
 
 public class SignalBUELight extends Signal {
@@ -11,6 +12,6 @@ public class SignalBUELight extends Signal {
 	}
 
 	public static final SEProperty<BUE_LIGHT> BUELIGHT = SEProperty.of("buelight", BUE_LIGHT.OFF);
-	public static final SEProperty<Boolean> NE2_2 = SEProperty.of("ne2_2", false);
-	public static final SEProperty<Boolean> NE2_4 = SEProperty.of("ne2_4", false);
+	public static final SEProperty<Boolean> NE2_2 = SEProperty.of("ne2_2", false, ChangeableStage.GUISTAGE);
+	public static final SEProperty<Boolean> NE2_4 = SEProperty.of("ne2_4", false, ChangeableStage.GUISTAGE);
 }